*: simplify data model
Now subscribes to each and every topic, storing all in memory for
now (let's see how far this scales).

Switches to regexp (instead of mqtt wildcards) for matching topics, as
Go has no nice and usable implementation of the latter.

Introduces a new 'regex' param for passing another regular expression
for filtering the message payload. When unnamed, the first capture group
is used. When named, the name is appended to the topic string and set to
the matched value.